Update Roadmap
Updated with new detminations regarding overall project scope. Creating a brand new GUI with Tkinter is currently outside of the realm of this class and I do not have the free time to learn it from scratch. I will proceed with the project as code without a GUI at this time.