Problems with Fixed Point Decimals and coversion to Binary
I have learned that languages have their own strengths and weaknesses, but it seems all of them share this weakness of floating point integers. Unfortunately, I will likely be unable to calculate the involved coefficients using Python due to Python's need to use numbers reprsented in Binary. This would not be a problem, execept I cannot represent decimals or irrational numbers using this method, both of which are required. I could use COBOL, since it defaults to BCD. Instead of a Wilk's calculator, this will be a new computerized method of calculating how well a lifter did. Think of it as Wilk's 3.0.