Numpy attempts
Recored failed code to date. Installed numpy and played with arrays. Attempted to get them to work for the means of this project but calculations came out incorrect just as before. Will begin looking into PySci, Panda's, ect. to see if they will be more usefule. The main problem now though still appears to be that Python calculations not based on binary are innacurate when using decimals and complex numbers.